keji_backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-socketio
python-socketio
eventlet
python-dotenv
psycopg2-binary
werkzeug
apscheduler
gunicorn
openai
tiktoken
resend
requests
cloudinary
certifi
